backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
pinecone
fastembed
langchain-text-splitters
pymupdf
python-multipart
python-dotenv
httpx
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
networkx
spacy
